start filling out full data model, resources and test cases

+++ src/main/java/cloudos/launcher/dao/LaunchConfigDAO.java
@@ -1,16 +1,14 @@
 package cloudos.launcher.dao;
 
+import cloudos.launcher.model.LaunchAccount;
 import cloudos.launcher.model.LaunchConfig;
-import cloudos.launcher.server.LaunchApiConfiguration;
-import lombok.Cleanup;
-import org.apache.commons.compress.utils.IOUtils;
-import org.cobbzilla.util.io.FileUtil;
-import org.cobbzilla.util.string.Base64;
 import org.cobbzilla.wizard.dao.UniquelyNamedEntityDAO;
+import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Repository;
 
 import javax.validation.Valid;
-import java.io.*;
+import java.io.File;
+import java.util.List;
 
 import static org.cobbzilla.util.daemon.ZillaRuntime.die;
 import static org.cobbzilla.util.io.FileUtil.abs;
@@ -18,63 +16,47 @@ import static org.cobbzilla.util.io.FileUtil.abs;
 @Repository
 public class LaunchConfigDAO extends UniquelyNamedEntityDAO<LaunchConfig> {
 
-    public static final String ZIP_FILE_DIR = "configs";
-
-    protected File getZipFileDir () { return LaunchApiConfiguration.configDir(ZIP_FILE_DIR); }
-
-    private File getZipFile(String uuid) {
-        return new File(getZipFileDir(), "cloudos-launch-config-"+uuid+".zip");
-    }
-
-    public String readZipData(String uuid) {
-        final File zipFile = getZipFile(uuid);
-        try {
-            @Cleanup final InputStream in = new FileInputStream(zipFile);
-            @Cleanup final ByteArrayOutputStream out = new ByteArrayOutputStream((int) (zipFile.length() + (zipFile.length() / 3)));
-            IOUtils.copy(in, out);
-            return Base64.encodeBytes(out.toByteArray());
-
-        } catch (Exception e) {
-            die("readZipData: "+e, e);
-        }
-        return FileUtil.toStringOrDie(zipFile);
-    }
-
-    protected void writeZipData(@Valid LaunchConfig entity) {
-        if (!entity.hasBase64zipData()) die("writeZipData: no base64zipData provided");
-        final File zipFile = getZipFile(entity.getUuid());
-        try {
-            @Cleanup final InputStream in = new ByteArrayInputStream(Base64.decode(entity.getBase64zipData()));
-            @Cleanup final OutputStream out = new FileOutputStream(zipFile);
-            IOUtils.copy(in, out);
-        } catch (Exception e) {
-            die("preCreate: error writing base64zipData to disk: "+e, e);
-        }
-    }
-
     @Override
     public LaunchConfig findByName(String name) {
         final LaunchConfig found = super.findByName(name);
         if (found == null) return null;
-        return found.setBase64zipData(readZipData(found.getUuid()));
+        return found.readZipData();
     }
 
     @Override
-    public Object preCreate(@Valid LaunchConfig entity) {
-        writeZipData(entity);
-        return super.preCreate(entity);
+    public Object preCreate(@Valid LaunchConfig config) {
+        config.writeZipData();
+        return super.preCreate(config);
     }
 
     @Override
-    public Object preUpdate(@Valid LaunchConfig entity) {
-        writeZipData(entity);
-        return super.preUpdate(entity);
+    public Object preUpdate(@Valid LaunchConfig config) {
+        config.writeZipData();
+        return super.preUpdate(config);
     }
 
     @Override
     public void delete(String uuid) {
-        final File zipFile = getZipFile(uuid);
+        final LaunchConfig config = findByUuid(uuid);
+        if (config == null) die("not found: "+uuid);
+        final File zipFile = config.getZipFile();
         if (!zipFile.delete()) die("Error deleting zipFile: "+abs(zipFile));
         super.delete(uuid);
+    }
+
+    public List<LaunchConfig> findByAccount(LaunchAccount account) {
+        final List<LaunchConfig> configs = findByField("account", account.getUuid());
+        for (LaunchConfig c : configs) {
+            c.setLaunchAccount(account).readZipData();
+        }
+        return configs;
+    }
+
+    public LaunchConfig findByAccountAndName(LaunchAccount account, String name) {
+        final LaunchConfig config = uniqueResult(criteria().add(
+                Restrictions.and(
+                        Restrictions.eq("account", account.getUuid()),
+                        Restrictions.eq("name", nameValue(name)))));
+        return config == null ? null : config.setLaunchAccount(account).readZipData();
     }
 }
